+++ docs/effect設置格式.lua
@@ -1,6 +1,6 @@
 setting = {
-    name,
+    (要跟檔名一樣)name,
     target,
-    value{percent, number},
-    time,
+    (可選)value{percent, number},
+    (可選)time,
 }