+++ src/data/effect/public/init.lua
@@ -1,0 +1,3 @@
+return require 'data.load_file'("data.effect.public.", {
+    "test",
+})