+++ src/war3/effect.lua
@@ -12,15 +12,16 @@ function Effect:_new(setting, manager)
         _class_ = setting.class,
         _value_ = setting.value,
         _level_ = 1,
-        _period_ = setting.period,
+        _period_ = setting.period or setting.time,
         _time_ = setting.time,
-        _priority_ = setting.priority,
+        _priority_ = setting.priority or 0,
         _mode_ = setting.mode,
+        _icon_ = setting.icon,
         _model_ = setting.model,
         _model_point_ = setting.model_point,
         _global_ = setting.global,
         _max_ = setting.max,
-        _keep_after_death_ = setting.keep_after_death,
+        _keep_after_death_ = setting.keep_after_death or false,
         _manager_ = manager,
         _tasks_ = require 'std.list':new(),
         on_add = setting.on_add,
@@ -31,14 +32,20 @@ function Effect:_new(setting, manager)
     }
 end
 
+function Effect:getName()
+    return self._name_
+end
+
 function Effect:start(new_task)
     InitVariables(new_task)
 
     if AddTask(self, new_task) then
         self:on_add(new_task)
-        AddModel(self, new_task)
+        -- AddModel(self, new_task)
         StartTimer(self, new_task)
     end
+
+    return self
 end
 
 InitVariables = function(task)
@@ -58,17 +65,16 @@ AddTask = function(self, new_task)
 
         -- 終止舊的效果
         self:delete(self._tasks_:front())
-
         -- 替換成新的效果
         self._tasks_:pop_front()
         self._tasks_:push_back(new_task)
         return true
     else -- 共存模式
         -- 比較優先級，新任務較高就覆蓋，都沒有就插入末端
-        for i, task in self._tasks_:iterator() do
-            if self.on_cover and self.on_cover(task, new_task) then
-                self._tasks_:insert(task, new_task)
-                CheckValid(self, i + 1, task)
+        for i, node in self._tasks_:iterator() do
+            if self.on_cover and self.on_cover(node:getData(), new_task) then
+                self._tasks_:insert(node, new_task)
+                CheckValid(self, i + 1, node:getData())
                 return CheckValid(self, i, new_task)
             end
         end
@@ -111,8 +117,9 @@ end
 
 function Effect:resume(task)
     self:on_add(task)
-    AddModel(self, task)
+    -- AddModel(self, task)
     task.timer:resume()
+    return self
 end
 
 AddModel = function(self, task)
@@ -125,20 +132,24 @@ function Effect:clear()
     for i = #self, 1, -1 do
         self:delete(self[i])
     end
+    
+    return self
 end
 
 function Effect:delete(task)
     task.timer:stop()
     self:on_delete(task)
-    DeleteModel(self, task)
+    -- DeleteModel(self, task)
     DeleteTask(self, task)
+    return self
 end
 
 function Effect:finish(task)
     self:on_finish(task)
     self:on_delete(task)
-    DeleteModel(self, task)
+    -- DeleteModel(self, task)
     DeleteTask(self, task)
+    return self
 end
 
 DeleteTask = function(self, task)
@@ -160,7 +171,8 @@ end
 function Effect:pause(task)
     task.timer:pause()
     self:on_delete(task)
-    DeleteModel(self, task)
+    -- DeleteModel(self, task)
+    return self
 end
 
 DeleteModel = function(self, task)